Attempting to implement a search icon input bar that opens and closes back to an icon.
